chore(webui): 契約テストをschemas配下へ寄せ深いimport制限に合わせる
